test: move snapshots into test folder
