--- src/tool/core/helpers/problem_metadata/tests.rs
@@ -1,44 +1,0 @@
-use std::io::Write as _;
-
-use anyhow::Context;
-use rstest::rstest;
-
-use crate::tool::core::helpers::{
-    local_store::{
-        path_local_store_problem_metadata,
-        tests::{get_rnd_request_delay, title_slugs, SlugList},
-    },
-    problem_metadata::{external_request_problem_metadata, get_problem_metadata},
-};
-
-#[rstest]
-#[ignore = "Only use for downloading responses"]
-fn download_response_from_leetcode(title_slugs: SlugList) {
-    for title_slug in title_slugs {
-        let sleep_delay = std::time::Duration::from_millis(get_rnd_request_delay());
-        println!(
-            "Going to sleep for {} milliseconds before requesting and trying to save {title_slug}",
-            sleep_delay.as_millis()
-        );
-        std::thread::sleep(sleep_delay); // Sleep to not go too hard on leetcode API
-        let response_string = external_request_problem_metadata(title_slug).unwrap();
-        let path = path_local_store_problem_metadata(title_slug);
-        let mut file = std::fs::OpenOptions::new()
-            .write(true)
-            .create(true)
-            .truncate(true)
-            .open(&path)
-            .with_context(|| format!("failed to save response to {path:?}"))
-            .unwrap();
-        file.write_all(response_string.as_bytes()).unwrap();
-        println!("Save of {title_slug:?} completed\n");
-    }
-    println!("Successfully saved all responses");
-}
-
-#[rstest]
-fn conversion_from_leetcode_response(title_slugs: SlugList) {
-    for title_slug in title_slugs {
-        insta::assert_debug_snapshot!(get_problem_metadata(title_slug).unwrap());
-    }
-}
